move around scripts to have more appropriate locations
