+++ Assets/MixedRealityToolkit.Extensions/MarkerDetection.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 1e1801e96223fff4d8963f02a7c882d8
+guid: 1fc142f7b8491cb4ab9deefb7313198e
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
