+++ Assets/MixedRealityToolkit.Extensions/PhotoCapture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b093de0651eec6742a96b019ebab27ce
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
