+++ Assets/MixedRealityToolkit.Extensions/PhotoCapture/HoloLensCamera.cs.meta
