+++ Assets/MixedRealityToolkit.Extensions/ScreenRecording/AndroidRecordingService.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
 using UnityEngine;
-using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Interfaces;
 using System;
 
 namespace Microsoft.MixedReality.Toolkit.Extensions.ScreenRecording
