+++ Assets/MixedRealityToolkit.Extensions/ScreenRecording/IRecordingServiceVisual.cs
@@ -1,9 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-using Microsoft.MixedReality.Toolkit.Extensions.ScreenRecording;
-
-namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Interfaces
+namespace Microsoft.MixedReality.Toolkit.Extensions.ScreenRecording
 {
     public interface IRecordingServiceVisual
     {
