+++ Assets/MixedRealityToolkit.Extensions/ScreenRecording/IRecordingServiceVisual.cs.meta
