+++ Assets/MixedRealityToolkit.Extensions/Sharing.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b67e122f2723f2242874c154bdb5671b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
