+++ Assets/MixedRealityToolkit.Extensions/Sharing/INetworkingService.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License. See LICENSE in the project root for license information.
 
-namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Interfaces
+namespace Microsoft.MixedReality.Toolkit.Extensions.Sharing
 {
     public enum NetworkPriority
     {
