+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/SpatialCoordinates/MarkerSpatialCoordinateService.cs
@@ -8,7 +8,8 @@ using UnityEngine;
 
 using UnityEngine.XR.ARFoundation;
 
-using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Interfaces;
+using Microsoft.MixedReality.Toolkit.Extensions.MarkerDetection;
+using Microsoft.MixedReality.Toolkit.Extensions.Sharing;
 using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Utilities;
 
 namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.SpatialCoordinates
