+++ Assets/MixedRealityToolkit.Extensions/SpectatorView/Scripts/SpectatorView.cs
@@ -4,9 +4,9 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Interfaces;
-using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Utilities;
 using Microsoft.MixedReality.Toolkit.Extensions.ScreenRecording;
+using Microsoft.MixedReality.Toolkit.Extensions.Sharing;
+using Microsoft.MixedReality.Toolkit.Extensions.SpectatorView.Utilities;
 
 namespace Microsoft.MixedReality.Toolkit.Extensions.SpectatorView
 {
